+++ packages/form-components/src/item-select/item-select.stories.ts
@@ -1,7 +1,6 @@
 import type { Meta, StoryFn } from "@storybook/react-vite";
 import h from "@macrostrat/hyper";
-import Box from "ui-box";
-import { ToasterContext } from "@macrostrat/ui-components";
+import { Box, ToasterContext } from "@macrostrat/ui-components";
 import { useState } from "react";
 import { ItemSelect } from ".";
 import {