+++ packages/ui-components/src/form-controls/color-editor.ts
@@ -1,9 +1,13 @@
-import h from "@macrostrat/hyper";
+import hyper from "@macrostrat/hyper";
 import { ErrorBoundary } from "../error-boundary";
 import { Sketch } from "@uiw/react-color";
 import { Button, Popover } from "@blueprintjs/core";
-import { useState } from "react";
 import chroma from "chroma-js";
+import { asChromaColor } from "@macrostrat/color-utils";
+import { useEffect, useMemo, useRef, useState } from "react";
+import styles from "./main.module.sass";
+
+const h = hyper.styled(styles);
 
 export function BasePopup(props) {
   const { children, isOpen, content, ...rest } = props;
@@ -68,5 +72,53 @@ export function ColorEditor(props) {
         },
       }),
     ],
+  );
+}
+
+export enum ColorConversionType {
+  HEX = "hex",
+  RGB = "rgb",
+  HSL = "hsl",
+  CSS = "css",
+}
+
+export function ColorPicker2({
+  value,
+  onChange,
+  editable = true,
+  type = ColorConversionType.CSS,
+}) {
+  const ref = useRef(null);
+  useEffect(() => {
+    if (ref.current == null) return;
+    ref.current.focus();
+  }, []);
+
+  const color = useMemo(() => {
+    try {
+      return asChromaColor(value).hex();
+    } catch {
+      return "#aaaaaa";
+    }
+  }, [value]);
+
+  return h(
+    "div.color-picker-container",
+    {
+      onKeyDown(evt) {
+        if (evt.key === "Escape") {
+          evt.preventDefault();
+        }
+      },
+      ref,
+      tabIndex: 0,
+    },
+    h(Sketch, {
+      color,
+      disableAlpha: true,
+      onChange(color) {
+        onChange(asChromaColor(color.hexa)[type]());
+      },
+    }),
   );
 }