+++ packages/ui-components/src/form-controls/index.ts
@@ -1,1 +1,2 @@
+export * from "./color-editor";
 export * from "./slider";