+++ packages/color-utils/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.1] - 2026-01-29
+
+- Change layout of `package.json` and explicitly mark no side effects.
+
 ## [1.2.0] - 2026-01-27
 
 Change build system from Parcel to Vite