+++ packages/column-components/CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file. The format
 is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this
 project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.4] - 2026-01-29
+
+- Change layout of `package.json`
+- Remove `ui-box` dependency in favor of `@macrostrat/ui-components`
+- Remove color picker
+- Bundle `labella` dependency to avoid CommonJS issues
+
 ## [2.0.3] - 2026-01-28
 
 Update handling of `@uiw/react-color` dependency