+++ packages/data-components/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [1.0.1] - 2026-01-29
+
+- Change layout of `package.json`
+
 ## [1.0.0] - 2026-01-26
 
 - Update peer dependencies to React 18